requirements.txt
Flask>=2.2.0
tensorflow-cpu>=2.10
numpy
h5py
Werkzeug
Pillow
python-dotenv
google-genai
